Update Xcode project using Xcode 9.3. Quiet/fix warnings about 64-to-32-bit integer sizes.

+++ lib/DDHidDevice.h
@@ -126,7 +126,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface DDHidDevice (Protected)
 
-- (NSInteger) sizeOfDefaultQueue;
+- (unsigned) sizeOfDefaultQueue;
 - (void) addElementsToDefaultQueue;
 
 @end
